docker report et gateway ok

+++ note-service/src/main/java/com/medilabosolutions/note_service/controller/NoteController.java
@@ -2,7 +2,9 @@ package com.medilabosolutions.note_service.controller;
 
 import com.medilabosolutions.note_service.dto.NoteCreateDTO;
 import com.medilabosolutions.note_service.dto.NoteDTO;
+import com.medilabosolutions.note_service.model.Note;
 import com.medilabosolutions.note_service.service.NoteService;
+import jakarta.annotation.PostConstruct;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -26,7 +28,7 @@ public class NoteController {
     }
 
     @GetMapping("/patient/{id}")
-    public ResponseEntity<List<NoteDTO>> getAllNotes(@PathVariable(name = "id") Long id) {
+    public ResponseEntity<List<NoteDTO>> getAllNotes(@PathVariable(name = "id") int id) {
         log.info("Getting all notes for: {}", id);
         return ResponseEntity.ok(noteService.getAllNotesByPatientId(id));
     }
