+++ note-service/src/main/java/com/medilabosolutions/note_service/dto/NoteCreateDTO.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record NoteCreateDTO(
         @NotNull
-        Long patId,
+        int patId,
 
         String patient,
