+++ note-service/src/main/java/com/medilabosolutions/note_service/dto/NoteDTO.java
@@ -2,7 +2,7 @@ package com.medilabosolutions.note_service.dto;
 
 public record NoteDTO(
         String id,
-        Long patId,
+        int patId,
         String patient,
         String note
 ) {
